requirements.txt
Flask>=3.1,<4
gunicorn>=23,<24
python-dotenv>=1.0,<2
PyJWT>=2.8,<3
Werkzeug>=3.1,<4
pandas>=2.2,<3
numpy>=2.1,<3
openpyxl>=3.1,<4
xlrd>=2.0,<3
psycopg2-binary>=2.9,<3
